requirements.txt
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
boto3==1.34.99
botocore==1.34.99
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
complexipy==0.3.0
coverage==7.5.0
cryptography==42.0.2
distlib==0.3.8
filelock==3.13.1
Flask==3.0.2
flaskwebgui==1.1.1
gdown==5.2.0
gunicorn==22.0.0
h11==0.14.0
idna==3.6
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.2.0
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pendulum==3.0.0
platformdirs==4.2.0
psutil==5.9.8
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.4
requests==2.31.0
requests-file==2.0.0
requests-toolbelt==1.0.0
rich==13.7.0
s3transfer==0.10.1
salesforce-bulk==2.2.0
selenium==4.17.2
shellingham==1.5.4
simple-salesforce==1.12.6
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
time-machine==2.13.0
tqdm==4.66.4
trio==0.24.0
trio-websocket==0.11.1
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
unicodecsv==0.14.1
urllib3==2.1.0
virtualenv==20.25.1
Werkzeug==3.0.1
wsproto==1.2.0
zeep==4.2.1
